requirements.txt
# Django Core
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Async Processing
celery==5.4.0
redis==5.0.4

# AWS Integration
boto3==1.34.131

# Environment Management
python-dotenv==1.0.1

# Development
django-extensions==3.2.3
ipython==8.24.0
django-filter==24.2
